backend/requirements.txt
# Backend Service Dependencies
paho-mqtt>=1.6.0
websockets>=10.0
